\name{ThresholdROC-package}
\alias{ThresholdROC-package}
\alias{ThresholdROC}
\docType{package}
\title{ Estimates the optimum threshold and its variance in diagnostic settings }
\description{ Estimates the optimum threshold and its variance in two and three state settings based on minimizing a cost function}
\details{
\tabular{ll}{
Package: \tab ThresholdROC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-07\cr
License: \tab GPL>=2\cr
LazyLoad: \tab yes\cr
}
}
\author{ Konstantina Skaltsa

Maintainer: Konstantina Skaltsa <kskaltsa@ub.edu> }
\references{
Skaltsa K et al. (2010) Estimation of the diagnostic threshold accounting for decision costs and sampling uncertainty. Biometrical Journal 52(5), 676-697.

Skaltsa K et al. (2011) Optimum threshold estimation based on cost function in a multistate diagnostic setting. Statistics in Medicine. In press.}
\keyword{ package }
\seealso{
}
\examples{
}
